Update package version and add PrimalRendDistance config
Updated RebornRotations.csproj to change the version of the
RotationSolverReborn.Basic package from 7.0.5.63 to 7.0.5.65.

Added a new configuration property PrimalRendDistance in
WAR_Default.cs, allowing the maximum distance for using the
Primal Rend ability to be set between 1 and 20 yalms.

Modified the logic for using the Primal Rend ability to utilize
the new PrimalRendDistance property instead of a hardcoded
value of 2 yalms.