Dancer Optimalizations
Shouldn't use abilities that their animation lock can drift dancing right before dance and during it, Just like Curing Waltz

Sometimes it used FanDance II or FanDance when FanDance III was available, now it doesn't.
However it hard to test since it happens rarely but i didn't notice a single instance of it happening for 2 hours so its probably working as intented

The overcapping mechanic for feathers now uses feathers before nextGCD that can generate it, instead of instantly when it hits 4, also moved the storing feathers mechanic to when the player hits level anough for devilment instead of tech step.

It seems that the code that prevented using standard step when technical step cd is less that time required to do the dance didn't work properly. Fixed that